+++ packages/shared/src/index.ts
@@ -13,3 +13,5 @@ export * from './request'
 export * from './directives'
 
 export * from './merge'
+
+export * from './imageResize'