+++ packages/sidenav/src/SideNav.vue
@@ -1,14 +1,14 @@
 <template>
-  <ScrollContainer :size="7" v-bind="getScrollContainerProps()" auto-hide>
-    <ul :class="showMini ? styles.sidenavMiniClass : styles.sidenavClass" v-bind="{ ...getContainerProps() }">
-      <slot name="menuTop"></slot>
-      <div v-show="search && !showMini" :class="styles.sidenavInputWrap">
-        <input v-bind="{ ...getSidenavInputProps() }" placeholder="search" :class="styles.sidenavInput" />
-      </div>
+  <!-- <ScrollContainer :size="7" v-bind="getScrollContainerProps()" auto-hide> -->
+  <ul :class="showMini ? styles.sidenavMiniClass : styles.sidenavClass" v-bind="{ ...getContainerProps() }">
+    <slot name="menuTop"></slot>
+    <div v-show="search && !showMini" :class="styles.sidenavInputWrap">
+      <input v-bind="{ ...getSidenavInputProps() }" placeholder="search" :class="styles.sidenavInput" />
+    </div>
 
-      <SideNode v-for="node in filteredMenu" :key="node[activeKey]" :node="node" :active-key="activeKey"
-        :key-field="keyField" :mini="showMini" :select-handler="selectNode">
-        <template #default="{
+    <SideNode v-for="node in filteredMenu" :key="node[activeKey]" :node="node" :active-key="activeKey"
+      :key-field="keyField" :mini="showMini" :select-handler="selectNode">
+      <template #default="{
           mini: miniValue,
           activeParent: activeParentValue,
           expand: expandValue,
@@ -16,20 +16,20 @@
           onClick: clickHandler,
           node: nodeValue,
         }">
-          <slot :mini="miniValue" :activeParent="activeParentValue" :expand="expandValue" :activeItem="activeItemValue"
-            :onClick="clickHandler" :node="nodeValue"></slot>
-        </template>
-      </SideNode>
+        <slot :mini="miniValue" :activeParent="activeParentValue" :expand="expandValue" :activeItem="activeItemValue"
+          :onClick="clickHandler" :node="nodeValue"></slot>
+      </template>
+    </SideNode>
 
-      <slot name="menuBottom"></slot>
-    </ul>
-  </ScrollContainer>
+    <slot name="menuBottom"></slot>
+  </ul>
+  <!-- </ScrollContainer> -->
 </template>
 
 <script setup lang="ts">
 import { withDefaults } from 'vue'
 import { style } from '@apathia/apathia.twind'
-import { ScrollContainer } from '@apathia/apathia.scroll-container'
+// import { ScrollContainer } from '@apathia/apathia.scroll-container'
 import SideNode from './SideNode.vue'
 import type { SideNode as Node } from './types'
 import useSideNav from './useSideNav'
@@ -61,15 +61,15 @@ const props = withDefaults(defineProps<SideNavProps>(), {
   keyField: 'key'
 })
 
-const emit = defineEmits(['select'])
+const emit = defineEmits(['select', 'minChange'])
 
 const {
   getContainerProps,
   getSidenavInputProps,
   filteredMenu,
   showMini,
-  getScrollContainerProps,
-} = useSideNav(props)
+  // getScrollContainerProps,
+} = useSideNav(props, emit)
 const styles = initStyle()
 
 const selectNode = (node: Node) => emit('select', node)