+++ packages/sidenav/src/types.ts
@@ -1,6 +1,8 @@
+import { Component, ComputedRef, Ref, ShallowRef } from 'vue'
+
 export type SideNode = {
   text?: string
-  icon?: string[]
+  icon?: Component
   children?: SideNode[]
   [k: string]: any
 }
@@ -11,4 +13,16 @@ export interface SideNavProps {
   mini?: boolean
   search?: boolean
   keyField?: keyof SideNode
+}
+
+export interface ISideNavReturn {
+  getContainerProps: () => {
+    onMouseenter: () => void
+    onMouseleave: () => void
+  }
+  getSidenavInputProps: () => {
+    onInput: (event: InputEvent) => void
+  }
+  filteredMenu: Ref<SideNode[]>
+  showMini: ComputedRef<boolean>
 }